requirements.txt
Flask==3.0.0
Pillow==10.4.0
Werkzeug==3.0.0